requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
greenlet==1.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
pandas==1.2.3
plotly==4.14.3
progressbar2==3.53.1
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
retrying==1.3.3
six==1.15.0
urllib3==1.26.4
Werkzeug==1.0.1
